+++ modules/app-desktop/src/main/kotlin/org/archivekeep/app/desktop/ui/views/home/components/SecondaryArchiveRepositoryRow.kt
@@ -66,6 +66,7 @@ fun SecondaryArchiveRepositoryRow(
             )
             InArchiveRepositoryDropdownIconLaunched(
                 repository = repository,
+                canAdd = nonPrimaryRepository.canAdd,
                 isAssociated = storageRepo.otherRepositoryState.associationId != null,
             )
         },