+++ modules/files/src/main/kotlin/org/archivekeep/utils/texts.kt
@@ -1,7 +1,9 @@
 package org.archivekeep.utils
 
-fun filesAutoPlural(items: Collection<*>): String =
-    if (items.size == 1) {
+fun filesAutoPlural(items: Collection<*>): String = filesAutoPlural(items.size)
+
+fun filesAutoPlural(count: Int): String =
+    if (count == 1) {
         "file"
     } else {
         "files"